remove test code for non-existing funcs
